Move user designation editing to `Users/Edit` page
Previously, users could edit their own designations (Program, Unit, Position) on the `Account/Index` page. This change consolidates that functionality into the `Users/Edit` page, allowing administrators to manage a user's roles and designations from a single administrative interface.

The `Account/Index` page is now a simplified, read-only display of the current user's profile, and the `Users/Details` page also consistently displays full user designation information. This improves administrative workflow and clarifies page responsibilities.
