+++ FMS/Pages/Account/Index.cshtml.cs
@@ -1,46 +1,18 @@
 using FMS.Domain.Services;
-using FMS.Platform.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using System.ComponentModel.DataAnnotations;
 
 namespace FMS.Pages.Account
 {
     public class IndexModel : PageModel
     {
         private readonly IUserService _userService;
-        private readonly ISelectListHelper _listHelper;
-        public IndexModel(IUserService userService, ISelectListHelper listHelper)
-        {
-            _userService = userService;
-            _listHelper = listHelper;
-        }
+        public IndexModel(IUserService userService) => _userService = userService;
 
         [BindProperty]
         public Guid Id { get; set; }
         public UserView CurrentUser { get; private set; }
         public IList<string> Roles { get; private set; }
-
-        [BindProperty]
-        [Display(Name = "Program")]
-        public Guid? UserProgramId { get; set; }
-
-        [BindProperty]
-        [Display(Name = "Unit")]
-        public Guid? UserUnitId { get; set; }
-        
-        [BindProperty]
-        [Display(Name = "Position")]
-        public Guid? UserPositionId { get; set; }
-
-        public SelectList UserUnits { get; private set; } 
-
-        public SelectList UserPrograms { get; private set; }
-
-        public SelectList UserPositions { get; private set; }
-
-        public DisplayMessage Message { get; set; }
 
         public async Task<IActionResult> OnGetAsync(Guid? id)
         {
@@ -57,53 +29,7 @@ namespace FMS.Pages.Account
                 Id = CurrentUser.Id;
             }
             Roles = await _userService.GetCurrentUserRolesAsync();
-            UserProgramId = CurrentUser.UserProgram?.Id;
-            UserUnitId = CurrentUser.UserUnit?.Id;
-            UserPositionId = CurrentUser.UserPosition?.Id;
-            await PopulateSelectsAsync();
-            Message = TempData?.GetDisplayMessage();
             return Page();
-        }
-
-        public async Task<ActionResult> OnPostAsync()
-        {
-            if (Id != Guid.Empty)
-            {
-                CurrentUser = await _userService.GetUserByIdAsync(Id)
-                 ?? throw new Exception("User not found");
-            }
-            else
-            {
-                TempData?.SetDisplayMessage(Context.Danger, $"Unable to find user with Id: {Id.ToString()}");
-                await PopulateSelectsAsync();
-                return Page();
-            }
-           
-            Roles = await _userService.GetCurrentUserRolesAsync();
-            try
-            {
-                await _userService.UpdateUserDesignationsAsync(CurrentUser.Id, 
-                    UserProgramId,
-                    UserUnitId,
-                    UserPositionId);
-            }
-            catch (Exception ex)
-            {
-                TempData?.SetDisplayMessage(Context.Danger, $"Unable to update user designations: {ex.Message}");
-                await PopulateSelectsAsync();
-                return Page();
-            }
-
-            TempData?.SetDisplayMessage(Context.Success, $"User designations successfully updated.");
-            return RedirectToPage();
-        }   
-
-        private async Task PopulateSelectsAsync()
-        {
-            UserPrograms = await _listHelper.UserProgramsSelectListAsync();
-            UserUnits = await _listHelper.OrganizationalUnitsSelectListAsync();
-            UserPositions = await _listHelper.UserPositionsSelectListAsync();
-            // false, ["Remedial Sites 1", "Remedial Sites 2", "Remedial Sites 3", "DOD Facilities", "NPL Unit", "Treatment & Storage", "SW Env. Monitoring Compliance", "Voluntary Remediation"]
         }
     }
 }
