+++ FMS/Pages/Users/Details.cshtml.cs
@@ -11,6 +11,9 @@ namespace FMS.Pages.Users
         public DetailsModel(IUserService userService) => _userService = userService;
 
         public string Id { get; private set; }
+
+        public UserView CurrentUser { get; private set; }
+
         public string DisplayName { get; private set; }
         public string Email { get; private set; }
         public IList<string> Roles { get; private set; }
@@ -18,21 +21,22 @@ namespace FMS.Pages.Users
 
         public async Task<IActionResult> OnGetAsync(Guid? id)
         {
-            if (id == null)
+            if (id != null)
             {
-                return NotFound();
+                CurrentUser = await _userService.GetUserByIdAsync(id.Value)
+                 ?? throw new Exception("User not found");
+                Id = CurrentUser.Id.ToString();
+            }
+            else
+            {
+                CurrentUser = await _userService.GetCurrentUserAsync()
+                    ?? throw new Exception("Current user not found");
+                Id = CurrentUser.Id.ToString();
             }
 
-            var user = await _userService.GetUserByIdAsync(id.Value);
-            if (user == null)
-            {
-                return NotFound();
-            }
-
-            Id = user.Id.ToString();
-            DisplayName = user.DisplayName;
-            Email = user.Email;
-            Roles = await _userService.GetUserRolesAsync(user.Id);
+            DisplayName = CurrentUser.DisplayName;
+            Email = CurrentUser.Email;
+            Roles = await _userService.GetUserRolesAsync(CurrentUser.Id);
 
             Message = TempData?.GetDisplayMessage();
             return Page();
