+++ FMS/Pages/Users/Edit.cshtml.cs
@@ -4,6 +4,8 @@ using FMS.Platform.Extensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using System.ComponentModel.DataAnnotations;
 
 namespace FMS.Pages.Users
 {
@@ -11,7 +13,12 @@ namespace FMS.Pages.Users
     public class EditModel : PageModel
     {
         private readonly IUserService _userService;
-        public EditModel(IUserService userService) => _userService = userService;
+        private readonly ISelectListHelper _listHelper;
+        public EditModel(IUserService userService, ISelectListHelper listHelper)
+        {
+            _userService = userService;
+            _listHelper = listHelper;
+        }
 
         [BindProperty]
         [HiddenInput]
@@ -35,25 +42,90 @@ namespace FMS.Pages.Users
         public string DisplayName { get; private set; }
         public string Email { get; private set; }
 
+        [BindProperty]
+        public Guid Id { get; set; }
+        public UserView CurrentUser { get; private set; }
+        public IList<string> Roles { get; private set; }
+
+        [BindProperty]
+        [Display(Name = "Program")]
+        public Guid? UserProgramId { get; set; }
+
+        [BindProperty]
+        [Display(Name = "Unit")]
+        public Guid? UserUnitId { get; set; }
+
+        [BindProperty]
+        [Display(Name = "Position")]
+        public Guid? UserPositionId { get; set; }
+
+        public SelectList UserUnits { get; private set; }
+
+        public SelectList UserPrograms { get; private set; }
+
+        public SelectList UserPositions { get; private set; }
+
+        public DisplayMessage Message { get; set; }
+
         public async Task<IActionResult> OnGetAsync(Guid? id)
         {
-            if (id == null)
+            if (id != null)
             {
-                return NotFound();
+                CurrentUser = await _userService.GetUserByIdAsync(id.Value)
+                 ?? throw new Exception("User not found");
+                Id = CurrentUser.Id;
+            }
+            else
+            {
+                CurrentUser = await _userService.GetCurrentUserAsync()
+                    ?? throw new Exception("Current user not found");
+                Id = CurrentUser.Id;
             }
 
             UserId = id.Value;
             if (!await GetUserDetails()) return NotFound();
             await GetUserRoles();
+
+            Roles = await _userService.GetCurrentUserRolesAsync();
+            UserProgramId = CurrentUser.UserProgram?.Id;
+            UserUnitId = CurrentUser.UserUnit?.Id;
+            UserPositionId = CurrentUser.UserPosition?.Id;
+            await PopulateSelectsAsync();
+            Message = TempData?.GetDisplayMessage();
+
             return Page();
         }
 
         public async Task<IActionResult> OnPostAsync()
         {
-            if (!ModelState.IsValid)
+            if (Id != Guid.Empty)
             {
+                CurrentUser = await _userService.GetUserByIdAsync(Id)
+                 ?? throw new Exception("User not found");
+            }
+            else
+            {
+                TempData?.SetDisplayMessage(Context.Danger, $"Unable to find user with Id: {Id.ToString()}");
+                await PopulateSelectsAsync();
                 return Page();
             }
+
+            Roles = await _userService.GetCurrentUserRolesAsync();
+            try
+            {
+                await _userService.UpdateUserDesignationsAsync(CurrentUser.Id,
+                    UserProgramId,
+                    UserUnitId,
+                    UserPositionId);
+            }
+            catch (Exception ex)
+            {
+                TempData?.SetDisplayMessage(Context.Danger, $"Unable to update user designations: {ex.Message}");
+                await PopulateSelectsAsync();
+                return Page();
+            }
+
+            TempData?.SetDisplayMessage(Context.Success, $"User designations successfully updated.");
 
             var roleSettings = new Dictionary<string, bool>()
             {
@@ -103,6 +175,14 @@ namespace FMS.Pages.Users
             HasFileCreatorRole = roles.Contains(UserRoles.FileCreator);
             HasFileEditorRole = roles.Contains(UserRoles.FileEditor);
             HasComplianceOfficerRole = roles.Contains(UserRoles.ComplianceOfficer);
+        }
+
+        private async Task PopulateSelectsAsync()
+        {
+            UserPrograms = await _listHelper.UserProgramsSelectListAsync();
+            UserUnits = await _listHelper.OrganizationalUnitsSelectListAsync();
+            UserPositions = await _listHelper.UserPositionsSelectListAsync();
+            // false, ["Remedial Sites 1", "Remedial Sites 2", "Remedial Sites 3", "DOD Facilities", "NPL Unit", "Treatment & Storage", "SW Env. Monitoring Compliance", "Voluntary Remediation"]
         }
     }
 }
